Rename mujoco_all_sac_lsp -> mujoco_all_sac
